aide prof matrix multiplication

+++ PD_project/matrix_multiplication_bench_2/cpu_functions.cpp
@@ -1,6 +1,6 @@
 #include "cpu_functions.h"
 
-typedef double bench_t;
+typedef int bench_t;
 
 void matrix_multiplication(const bench_t* A, const bench_t* B, bench_t* C,const unsigned int n, const unsigned int m, const unsigned int w ){
 	for (unsigned int i = 0; i < n; ++i)
